made separate gamepage component
